More horizontal line fixes. (#1423)
* More horizontal line fixes.

* Adding horizontal divider to separate post and comments in postactivity.

---------

+++ app/src/main/java/com/jerboa/Utils.kt
@@ -1226,7 +1226,7 @@ fun calculateCommentOffset(
     return if (depth == 0) {
         0.dp
     } else {
-        (abs((depth.minus(1) * padding.value.toInt())).dp + padding)
+        (abs((depth.minus(1) * padding.value)).dp + padding)
     }
 }
